add babel6, move testUtils to __tests__
