+++ packages/babel-plugin-jss/package.json
@@ -8,7 +8,8 @@
   "license": "MIT",
   "main": "./lib/index.js",
   "devDependencies": {
-    "@babel/core": "^7.0.0-beta.53",
+    "@babel/core": "^7.0.0-beta.54",
+    "babel-core": "^6.26.3",
     "babel-plugin-syntax-decorators": "^6.13.0",
     "jest": "^23.4.0"
   },
